+++ themes/base/joomla/html/layouts/chromes/gantry.php
@@ -30,14 +30,9 @@ $moduleTag              = $params->get('module_tag', 'div');
 $moduleAttribs          = [];
 $moduleAttribs['class'] = $module->position . ' moduletable ' . htmlspecialchars($params->get('moduleclass_sfx'), ENT_QUOTES, 'UTF-8');
 $headerTag              = htmlspecialchars($params->get('header_tag', 'h3'), ENT_QUOTES, 'UTF-8');
-$headerClass            = htmlspecialchars($params->get('header_class', ''), ENT_QUOTES, 'UTF-8');
+$headerClass            = htmlspecialchars($params->get('header_class', 'g-title'), ENT_QUOTES, 'UTF-8');
 $headerAttribs          = [];
 $headerAttribs['class'] = $headerClass;
-
-// Only output a header class if it is not card-title
-if ($headerClass !== 'card-title') :
-    $headerAttribs['class'] = 'card-header ' . $headerClass;
-endif;
 
 // Only add aria if the moduleTag is not a div
 if ($moduleTag !== 'div')
@@ -53,13 +48,8 @@ if ($moduleTag !== 'div')
 $header = '<' . $headerTag . ' ' . ArrayHelper::toString($headerAttribs) . '>' . $module->title . '</' . $headerTag . '>';
 ?>
 <<?php echo $moduleTag; ?> <?php echo ArrayHelper::toString($moduleAttribs); ?>>
-    <?php if ($module->showtitle && $headerClass !== 'card-title') : ?>
+    <?php if ($module->showtitle) : ?>
         <?php echo $header; ?>
     <?php endif; ?>
-    <div class="card-body">
-        <?php if ($module->showtitle && $headerClass === 'card-title') : ?>
-            <?php echo $header; ?>
-        <?php endif; ?>
-        <?php echo $module->content; ?>
-    </div>
+    <?php echo $module->content; ?>
 </<?php echo $moduleTag; ?>>